# OpenAPI/openapi-8.json
{"openapi":"3.1.0","info":{"title":"AeroLink Passenger Service","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/v1/passengers/profile":{"get":{"summary":"Get Profile","operationId":"get_profile_api_v1_passengers_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}},"security":[{"HTTPBearer":[]}]},"put":{"summary":"Update Profile","operationId":"update_profile_api_v1_passengers_profile_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Create Profile","operationId":"create_profile_api_v1_passengers_profile_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]},"delete":{"summary":"Delete Profile","description":"GDPR right to erasure - delete passenger profile and cascade to related services.","operationId":"delete_profile_api_v1_passengers_profile_delete","responses":{"204":{"description":"Successful Response"}},"security":[{"HTTPBearer":[]}]}},"/api/v1/passengers/consent":{"get":{"summary":"List Consents","description":"List all GDPR consents for the current user.","operationId":"list_consents_api_v1_passengers_consent_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ConsentResponse"},"type":"array","title":"Response List Consents Api V1 Passengers Consent Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Grant Consent","description":"Record GDPR consent for a specific data processing purpose.","operationId":"grant_consent_api_v1_passengers_consent_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"ConsentRequest":{"properties":{"purpose":{"type":"string","minLength":1,"title":"Purpose"},"granted":{"type":"boolean","title":"Granted"}},"type":"object","required":["purpose","granted"],"title":"ConsentRequest"},"ConsentResponse":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"purpose":{"type":"string","title":"Purpose"},"granted":{"type":"boolean","title":"Granted"},"granted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Granted At"},"revoked_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Revoked At"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","purpose","granted","granted_at","revoked_at","created_at"],"title":"ConsentResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","default":"healthy"},"service":{"type":"string","title":"Service"},"uptime_seconds":{"type":"number","title":"Uptime Seconds"},"dependencies":{"additionalProperties":{"type":"string"},"type":"object","title":"Dependencies","default":{}}},"type":"object","required":["service","uptime_seconds"],"title":"HealthResponse"},"ProfileCreate":{"properties":{"title":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Title"},"gender":{"anyOf":[{"type":"string","maxLength":20},{"type":"null"}],"title":"Gender"},"first_name":{"type":"string","maxLength":100,"minLength":1,"title":"First Name"},"middle_name":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Middle Name"},"last_name":{"type":"string","maxLength":100,"minLength":1,"title":"Last Name"},"date_of_birth":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Birth"},"nationality":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nationality"},"passport_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Passport Number"},"passport_expiry":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Passport Expiry"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"meal_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Meal Preference"},"seat_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Seat Preference"},"emergency_contact":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Emergency Contact"}},"type":"object","required":["first_name","last_name"],"title":"ProfileCreate"},"ProfileResponse":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender"},"first_name":{"type":"string","title":"First Name"},"middle_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Middle Name"},"last_name":{"type":"string","title":"Last Name"},"date_of_birth":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Birth"},"nationality":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nationality"},"passport_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Passport Number"},"passport_expiry":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Passport Expiry"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"loyalty_tier":{"type":"string","title":"Loyalty Tier"},"loyalty_points":{"type":"integer","title":"Loyalty Points"},"meal_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Meal Preference"},"seat_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Seat Preference"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","title","gender","first_name","middle_name","last_name","date_of_birth","nationality","passport_number","passport_expiry","phone_number","loyalty_tier","loyalty_points","meal_preference","seat_preference","created_at"],"title":"ProfileResponse"},"ProfileUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"gender":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gender"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"middle_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Middle Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"date_of_birth":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Birth"},"nationality":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nationality"},"passport_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Passport Number"},"passport_expiry":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Passport Expiry"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number"},"meal_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Meal Preference"},"seat_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Seat Preference"},"emergency_contact":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Emergency Contact"}},"type":"object","title":"ProfileUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}