# services/passenger-service/main.py
import sys
import os
import time

sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))

from datetime import datetime, timezone
from contextlib import asynccontextmanager
from fastapi import FastAPI, Depends, HTTPException, Request
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select, text
from shared.health import check_db, check_redis
import httpx
from shared.config import BaseConfig
from shared.database import create_db_engine, create_session_factory, Base
from fastapi.security import APIKeyHeader
from shared.auth import get_current_user
from shared.resilience import async_retry
from shared.encryption import encrypt_field, decrypt_field
from shared.audit import AuditLog, record_audit
from shared.cache import create_redis_client, redis_get_json, redis_set_json, redis_delete
from shared.logging import setup_logging
from shared.tracing import TraceMiddleware
from shared.schemas import HealthResponse
from models import PassengerProfile, GDPRConsent
from schemas import (ProfileCreate, ProfileUpdate, ProfileResponse,
                     ConsentRequest, ConsentResponse)

config = BaseConfig(service_name="passenger-service")
setup_logging(config.service_name)

engine = create_db_engine(config.database_url)
SessionFactory = create_session_factory(engine)
START_TIME = time.time()

redis_client = create_redis_client(config.redis_url)
PROFILE_CACHE_TTL = 300  # 5 minutes; invalidated on update or loyalty award

# Internal service-to-service API key — shared with payment-service via INTERNAL_API_KEY env var.
# This protects the loyalty/award endpoint from unauthenticated external callers.
_INTERNAL_API_KEY = os.environ.get("INTERNAL_API_KEY", "")
_internal_key_header = APIKeyHeader(name="X-Internal-API-Key", auto_error=False)

async def require_internal_key(key: str | None = Depends(_internal_key_header)):
    if not _INTERNAL_API_KEY or key != _INTERNAL_API_KEY:
        raise HTTPException(status_code=403, detail="Forbidden")


async def get_db():
    async with SessionFactory() as session:
        try:
            yield session
            await session.commit()
        except Exception:
            await session.rollback()
            raise


def _to_response(p: PassengerProfile) -> ProfileResponse:
    # Decrypt passport number on read
    passport = None
    if p.passport_number_encrypted:
        try:
            passport = decrypt_field(p.passport_number_encrypted)
        except Exception:
            passport = "***ENCRYPTED***"

    return ProfileResponse(
        id=str(p.id), user_id=str(p.user_id),
        title=p.title, gender=p.gender,
        first_name=p.first_name, middle_name=p.middle_name, last_name=p.last_name,
        date_of_birth=p.date_of_birth, nationality=p.nationality,
        passport_number=passport, passport_expiry=p.passport_expiry,
        phone_number=p.phone_number,
        loyalty_tier=p.loyalty_tier, loyalty_points=p.loyalty_points,
        meal_preference=p.meal_preference, seat_preference=p.seat_preference,
        created_at=p.created_at,
    )


@asynccontextmanager
async def lifespan(app: FastAPI):
    async with engine.begin() as conn:
        try:
            await conn.run_sync(Base.metadata.create_all)
        except Exception:
            pass  # Table may already exist from another service starting concurrently
        await conn.execute(text(
            "ALTER TABLE passenger_profiles ADD COLUMN IF NOT EXISTS middle_name VARCHAR(100)"
        ))
        await conn.execute(text(
            "ALTER TABLE passenger_profiles ALTER COLUMN nationality TYPE VARCHAR(100)"
        ))
        await conn.execute(text(
            "ALTER TABLE passenger_profiles ADD COLUMN IF NOT EXISTS title VARCHAR(20)"
        ))
        await conn.execute(text(
            "ALTER TABLE passenger_profiles ADD COLUMN IF NOT EXISTS gender VARCHAR(20)"
        ))
        await conn.execute(text(
            "ALTER TABLE passenger_profiles ADD COLUMN IF NOT EXISTS passport_expiry DATE"
        ))
    yield
    await engine.dispose()


app = FastAPI(title="AeroLink Passenger Service", version="1.0.0", lifespan=lifespan)
app.add_middleware(TraceMiddleware)


@app.get("/health", response_model=HealthResponse)
async def health(db: AsyncSession = Depends(get_db)):
    return HealthResponse(service="passenger-service", uptime_seconds=time.time() - START_TIME,
                          dependencies={"database": await check_db(db),
                                        "redis": check_redis(redis_client)})


@app.post("/api/v1/passengers/profile", response_model=ProfileResponse, status_code=201)
async def create_profile(data: ProfileCreate,
                         current_user: dict = Depends(get_current_user),
                         db: AsyncSession = Depends(get_db)):
    existing = await db.execute(
        select(PassengerProfile).where(PassengerProfile.user_id == current_user["sub"]))
    if existing.scalar_one_or_none():
        raise HTTPException(status_code=409, detail="Profile already exists")

    # Encrypt PII fields before storage
    encrypted_passport = encrypt_field(data.passport_number) if data.passport_number else None

    profile = PassengerProfile(
        user_id=current_user["sub"],
        title=data.title,
        gender=data.gender,
        first_name=data.first_name,
        middle_name=data.middle_name,
        last_name=data.last_name,
        date_of_birth=data.date_of_birth,
        nationality=data.nationality,
        passport_number_encrypted=encrypted_passport,
        passport_expiry=data.passport_expiry,
        phone_number=data.phone_number,
        meal_preference=data.meal_preference,
        seat_preference=data.seat_preference,
        emergency_contact=data.emergency_contact,
    )
    db.add(profile)
    await db.flush()
    await db.refresh(profile)
    response = _to_response(profile)
    redis_set_json(redis_client, f"profile:{current_user['sub']}", response.dict(), PROFILE_CACHE_TTL)
    return response


@app.get("/api/v1/passengers/profile", response_model=ProfileResponse)
async def get_profile(current_user: dict = Depends(get_current_user),
                      db: AsyncSession = Depends(get_db)):
    cache_key = f"profile:{current_user['sub']}"
    cached = redis_get_json(redis_client, cache_key)
    if cached is not None:
        return cached
    result = await db.execute(
        select(PassengerProfile).where(PassengerProfile.user_id == current_user["sub"]))
    profile = result.scalar_one_or_none()
    if not profile:
        raise HTTPException(status_code=404, detail="Profile not found")
    data = _to_response(profile).dict()
    redis_set_json(redis_client, cache_key, data, PROFILE_CACHE_TTL)
    return data


@app.put("/api/v1/passengers/profile", response_model=ProfileResponse)
async def update_profile(data: ProfileUpdate,
                         current_user: dict = Depends(get_current_user),
                         db: AsyncSession = Depends(get_db)):
    result = await db.execute(
        select(PassengerProfile).where(PassengerProfile.user_id == current_user["sub"]))
    profile = result.scalar_one_or_none()
    if not profile:
        raise HTTPException(status_code=404, detail="Profile not found")

    update_data = data.dict(exclude_unset=True)
    # Encrypt passport if being updated
    if "passport_number" in update_data:
        val = update_data.pop("passport_number")
        profile.passport_number_encrypted = encrypt_field(val) if val else None
    for key, value in update_data.items():
        setattr(profile, key, value)
    await db.flush()
    await db.refresh(profile)
    response = _to_response(profile)
    redis_set_json(redis_client, f"profile:{current_user['sub']}", response.dict(), PROFILE_CACHE_TTL)
    return response


TIER_THRESHOLDS = [
    (10_000, "platinum"),
    (5_000,  "gold"),
    (1_000,  "silver"),
    (0,      "bronze"),
]

def _tier_for(points: int) -> str:
    for threshold, tier in TIER_THRESHOLDS:
        if points >= threshold:
            return tier
    return "bronze"

POINTS_PER_DOLLAR = 10  # 10 pts per $1 spent


@app.post("/api/v1/passengers/loyalty/award")
async def award_loyalty_points(
    data: dict,
    _: None = Depends(require_internal_key),
    db: AsyncSession = Depends(get_db),
):
    """Internal endpoint — called by payment-service after successful payment."""
    user_id = data.get("user_id")
    points  = int(data.get("points", 0))
    if not user_id or points <= 0:
        raise HTTPException(status_code=400, detail="user_id and positive points required")

    result = await db.execute(
        select(PassengerProfile).where(PassengerProfile.user_id == user_id))
    profile = result.scalar_one_or_none()
    if not profile:
        raise HTTPException(status_code=404, detail="Profile not found")

    profile.loyalty_points = (profile.loyalty_points or 0) + points
    profile.loyalty_tier   = _tier_for(profile.loyalty_points)
    await db.flush()
    await db.refresh(profile)

    # Bust profile cache — loyalty data changed
    redis_delete(redis_client, f"profile:{user_id}")

    return {
        "user_id": user_id,
        "points_awarded": points,
        "total_points": profile.loyalty_points,
        "tier": profile.loyalty_tier,
    }


@app.delete("/api/v1/passengers/profile", status_code=204)
async def delete_profile(request: Request,
                         current_user: dict = Depends(get_current_user),
                         db: AsyncSession = Depends(get_db)):
    """GDPR right to erasure - delete passenger profile and cascade to related services."""
    result = await db.execute(
        select(PassengerProfile).where(PassengerProfile.user_id == current_user["sub"]))
    profile = result.scalar_one_or_none()
    if not profile:
        raise HTTPException(status_code=404, detail="Profile not found")

    # Delete consent records
    consents = await db.execute(
        select(GDPRConsent).where(GDPRConsent.user_id == current_user["sub"]))
    for consent in consents.scalars().all():
        await db.delete(consent)

    # Audit the deletion
    client_ip = request.client.host if request.client else None
    await record_audit(db, "passenger-service", "gdpr.erasure", "passenger_profile",
                       resource_id=str(profile.id), actor_id=current_user["sub"],
                       actor_role=current_user.get("role"),
                       detail="GDPR right to erasure exercised - profile and consents deleted",
                       ip_address=client_ip)

    # Cascade: request deletion from other services (GDPR right to erasure).
    # Failures are logged but do not block local deletion — a separate reconciliation
    # process should retry any cascades that did not return 2xx.
    for service_url, path in [
        (config.booking_service_url, f"/api/v1/bookings/user/{current_user['sub']}/anonymize"),
        (config.notification_service_url, f"/api/v1/notifications/user/{current_user['sub']}/anonymize"),
    ]:
        try:
            async for attempt in async_retry():
                with attempt:
                    async with httpx.AsyncClient(timeout=10.0) as client:
                        resp = await client.post(f"{service_url}{path}")
                        resp.raise_for_status()
        except Exception as e:
            logger.error("gdpr_cascade_failed",
                         service_url=service_url,
                         path=path,
                         user_id=current_user["sub"],
                         error=str(e))

    redis_delete(redis_client, f"profile:{current_user['sub']}")
    await db.delete(profile)


# ─── GDPR Consent Management ────────────────────────────────

@app.post("/api/v1/passengers/consent", response_model=ConsentResponse, status_code=201)
async def grant_consent(data: ConsentRequest, request: Request,
                        current_user: dict = Depends(get_current_user),
                        db: AsyncSession = Depends(get_db)):
    """Record GDPR consent for a specific data processing purpose."""
    # Check for existing consent for this purpose
    existing = await db.execute(
        select(GDPRConsent).where(
            GDPRConsent.user_id == current_user["sub"],
            GDPRConsent.purpose == data.purpose))
    consent = existing.scalar_one_or_none()

    client_ip = request.client.host if request.client else None

    if consent:
        consent.granted = data.granted
        if data.granted:
            consent.granted_at = datetime.now(timezone.utc)
            consent.revoked_at = None
        else:
            consent.revoked_at = datetime.now(timezone.utc)
        consent.ip_address = client_ip
    else:
        consent = GDPRConsent(
            user_id=current_user["sub"],
            purpose=data.purpose,
            granted=data.granted,
            granted_at=datetime.now(timezone.utc) if data.granted else None,
            ip_address=client_ip,
        )
        db.add(consent)

    await db.flush()
    await db.refresh(consent)

    await record_audit(db, "passenger-service",
                       f"consent.{'granted' if data.granted else 'revoked'}",
                       "gdpr_consent", resource_id=str(consent.id),
                       actor_id=current_user["sub"],
                       detail=f"Purpose: {data.purpose}", ip_address=client_ip)

    return ConsentResponse(
        id=str(consent.id), user_id=str(consent.user_id),
        purpose=consent.purpose, granted=consent.granted,
        granted_at=consent.granted_at, revoked_at=consent.revoked_at,
        created_at=consent.created_at,
    )


@app.get("/api/v1/passengers/consent", response_model=list[ConsentResponse])
async def list_consents(current_user: dict = Depends(get_current_user),
                        db: AsyncSession = Depends(get_db)):
    """List all GDPR consents for the current user."""
    result = await db.execute(
        select(GDPRConsent).where(GDPRConsent.user_id == current_user["sub"]))
    return [ConsentResponse(
        id=str(c.id), user_id=str(c.user_id), purpose=c.purpose,
        granted=c.granted, granted_at=c.granted_at, revoked_at=c.revoked_at,
        created_at=c.created_at,
    ) for c in result.scalars().all()]


if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8007)